Add PrepareForPlayr.ps1 EnablePowershellScripts.cmd  DisablePowershellScripts.cmd
* The EnablePowershellScripts.cmd and DisablePowershellScripts.cmd files make
it easy for a user to enable/disable powerscript execution, since this
is by default disabled for security reasons.

* The PrepareForPlayr.ps1 is a powershell script that will automatically
install/enable/configure Windows settings and applications to simplify
the procedure to prepare a device running Windows10 for use as a Bizplay
player. This is a first version, the goal is to increase the
fucntionality if possible to further reduce the manual steps that have
to be done to set up Bizplay playback on a device.

* Things learned;
- Although multiple sources on the inetrnet claim to enable you to
download a stand alone Chrome intaller (to allow fully automated install
of Chrome) the installer exe seems to either contain a unique key or via another
method ascertains that it was recently downloaded at which time the user
agreed with the Chrome terms of service because running the stand alone
installer will result in most cases in some kind of error.
- ways to set up automatic start and stopping of the device were not
found yet