Add fallback to Edge and IE
* Since the newest versions of Edge use Blinck and score equal to Chrome
on html5 compatibility it makes sense to fall back to Edge when Chrome
and Chromium cannot be found.

* As a Hail Mary escape/default option when none of the other browsers
can be found Internet Explorer is used. This is not a "real" solution
(it might very well give errors as Chrome command line parameters are
used) but it will either work on very old Windows systems or at least
give the user something to complain about so we can help out via the
usual support channels.