Add flags to prevent problems with browser-app on http playback
* Since Chrome v89 the default cookie settings/handling has changed,
meaning that the browser-app will not show content that needs cookies
that are not available in the iframe because of these changed settings.

* The settings could be changed in the GUI of Chrome but since v91 this
is no longer possible. the applied command line option
--disable-features=SameSiteByDefaultCookies,CookiesWithoutSameSiteMustBeSecure
is the current remedy

* from v94 onward this command line option is no longer supported so we
will have to revert to https playback when the content of the
browser-app requires cookies to work